+++ notify_pod_flutter/lib/notify_pod_flutter.dart
@@ -1,2 +1,3 @@
 export 'src/utils/get_device_info_model.dart';
 export 'src/utils/read_device_info.dart';
+export 'src/widgets/notification_bell.dart';