+++ notify_pod_flutter/CHANGELOG.md
@@ -1,3 +1,10 @@
+# Changelog NotifyPod Flutter
+
 ## 1.0.0
 
 - Initial version, created by Chivo
+
+## 1.0.1
+
+- Added NotificationBell widget for easy integration of notification indicators in app bars.
+- The widget dynamically updates the notification count using Firebase Cloud Messaging.